+++ src/routes/index.tsx
@@ -2,6 +2,7 @@ import Background from "../islands/Background.tsx"
 import Home from "../islands/Home.tsx"
 import Usage from "../islands/Usage.tsx"
 import Confettis from "../islands/Confettis.tsx"
+import Author from "../islands/Author.tsx"
 import Footer from "../islands/Footer.tsx"
 
 export default function HomePage() {
@@ -11,6 +12,7 @@ export default function HomePage() {
 			<Home />
 			<Usage />
 			<Confettis />
+			<Author />
 			<Footer />
 		</>
   	)